refactor: OmockRule 객체 생성하여 의존성 분리
